+++ src/notification/notification.ts
@@ -3,20 +3,22 @@ import {sendEmail} from './email';
 import {sendSMS} from './sms';
 import {playSound} from './sound';
 import {sendSlackMessage} from './slack';
-import sendPushoverNotification from './pushover';
+import {sendPushoverNotification} from './pushover';
 import {sendTelegramMessage} from './telegram';
 import {sendDiscordMessage} from './discord';
 
+const notifications = Config.notifications;
+
 export function sendNotification(cartUrl: string) {
-	if (Config.notifications.email.username && Config.notifications.email.password) {
+	if (notifications.email.username && notifications.email.password) {
 		sendEmail(cartUrl);
 	}
 
-	if (Config.notifications.slack.channel && Config.notifications.slack.token) {
+	if (notifications.slack.channel && notifications.slack.token) {
 		sendSlackMessage(cartUrl);
 	}
 
-	if (Config.notifications.telegram.botToken && Config.notifications.telegram.chatId) {
+	if (notifications.telegram.accessToken && notifications.telegram.chatId) {
 		sendTelegramMessage(cartUrl);
 	}
 
@@ -24,18 +26,18 @@ export function sendNotification(cartUrl: string) {
 		sendDiscordMessage(cartUrl);
 	}
 
-	if (Config.notifications.phone.number) {
-		const carrier = Config.notifications.phone.carrier.toLowerCase();
-		if (carrier && Config.notifications.phone.availableCarriers.has(carrier)) {
+	if (notifications.phone.number) {
+		const carrier = notifications.phone.carrier.toLowerCase();
+		if (carrier && notifications.phone.availableCarriers.has(carrier)) {
 			sendSMS(cartUrl);
 		}
 	}
 
-	if (Config.notifications.pushover.token && Config.notifications.pushover.user) {
+	if (notifications.pushover.token && notifications.pushover.user) {
 		sendPushoverNotification(cartUrl);
 	}
 
-	if (Config.notifications.playSound !== 'false') {
+	if (notifications.playSound === 'true') {
 		playSound();
 	}
 }